requirements.txt
# OpenGodOS 数字生命操作系统依赖

# 核心依赖
PyYAML>=6.0.1
Flask>=3.0.3
Flask-CORS>=5.0.0
Flask-SocketIO>=5.4.0
python-socketio>=5.11.2
eventlet>=0.37.1

# AI集成
openai>=1.30.1
tiktoken>=0.7.0

# 开发工具
pytest>=8.0.0
pytest-cov>=5.0.0
pylint>=3.1.0
black>=24.1.1
mypy>=1.8.0

# 工具库
python-dotenv>=1.0.1
requests>=2.31.0
numpy>=1.26.4

# 可选：高级功能
scikit-learn>=1.4.0
pandas>=2.2.0
matplotlib>=3.8.3